Ethernet configuration

+++ build-docker.sh
@@ -8,9 +8,10 @@ for tag in latest stage0 stage1 stage2; do
   docker build . -t $IMAGE:$tag --build-arg TAG=$tag --platform linux/amd64,linux/i386
 done
 
-for tag in latest stage2; do
-  for arch in amd64 i386; do
-    docker run --rm --platform linux/$arch \
+if [ "$1" = "--test" ]; then
+  for tag in latest stage2; do
+    for arch in amd64 i386; do
+      docker run --rm --platform linux/$arch \
         -v $PWD/test-cc:/tmp/test-cc \
         -v $PWD/lib/toybox:/tmp/lib/toybox \
         -v $PWD/tarballs:/src/tarballs \
@@ -46,8 +47,9 @@ for tag in latest stage2; do
 
             echo All tests passed!
         "
+    done
   done
-done
+fi
 
 for tag in latest stage0 stage1 stage2; do
   for arch in amd64 i386; do
